better organization for the test project
